move the Motor class from bapsf_motion.controllers to bapsf_motion.actors
