+++ src/bapsf_motion/actors/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["Motor"]
+
+from bapsf_motion.actors.motor_ import Motor
